test client-hc routing without internet requirement
